+++ Source/MixerButton.cpp
@@ -26,10 +26,10 @@ void MixerButton::paint(juce::Graphics& g)
 {
   constexpr int cornerRadius = 10;
   constexpr int sliderMargin = 70;
-  const float halfWidth      = static_cast<int>(getWidth()) * 0.5f;
-  const float halfHeight     = static_cast<int>(getHeight()) * 0.5f;
-  const float thirdHeight    = static_cast<int>(getHeight()) * 0.333333f;
-  float sliderWidth          = static_cast<int>(getWidth()) - sliderMargin;
+  const float halfWidth      = static_cast<float>(getWidth()) * 0.5f;
+  const float halfHeight     = static_cast<float>(getHeight()) * 0.5f;
+  const float thirdHeight    = static_cast<float>(getHeight()) * 0.333333f;
+  float sliderWidth          = static_cast<float>(getWidth()) - sliderMargin;
   auto bounds                = juce::Rectangle<int>(60, 0, static_cast<int>(sliderWidth), getHeight());
   // Background
   g.setGradientFill(juce::ColourGradient(juce::Colours::grey, halfWidth, halfHeight, juce::Colours::darkgrey, 0, 0, true));
@@ -38,12 +38,12 @@ void MixerButton::paint(juce::Graphics& g)
   // Labels
   g.setColour(juce::Colours::snow);
   g.setFont(labelFont_.withHeight(24.0f));
-  g.drawFittedText("dirtier", bounds.removeFromTop(thirdHeight), juce::Justification::centred, 1);
+  g.drawFittedText("dirtier", bounds.removeFromTop(static_cast<int>(thirdHeight)), juce::Justification::centred, 1);
 
   g.setFont(24.0f);
-  g.drawLine(70, thirdHeight, getWidth() - 20, getHeight() / 3, 1);
-  g.drawFittedText("dirty", bounds.removeFromTop(thirdHeight), juce::Justification::centred, 1);
-  g.drawLine(70, thirdHeight * 2, getWidth() - 20, thirdHeight * 2, 1);
+  g.drawLine(70, thirdHeight, static_cast<float>(getWidth() - 20), static_cast<float>(getHeight()) / 3.0f, 1);
+  g.drawFittedText("dirty", bounds.removeFromTop(static_cast<int>(thirdHeight)), juce::Justification::centred, 1);
+  g.drawLine(70, thirdHeight * 2, static_cast<float>(getWidth() - 20), thirdHeight * 2, 1);
   g.drawFittedText("clean", bounds, juce::Justification::centred, 1);
 
   auto barBounds =
@@ -93,13 +93,13 @@ void MixerButton::mouseUp(const juce::MouseEvent& event)
 void MixerButton::mapMouseToValue(const juce::Point<int>& mPoint)
 {
   constexpr float xMin = 70.0f;
-  const float xMax     = getWidth() - 20.0f;
+  const float xMax     = static_cast<float>(getWidth()) - 20.0f;
   constexpr auto yMin  = kHandleHeight / 2.0f;
-  const auto yMax      = getHeight() - (kHandleHeight / 2.0f);
+  const auto yMax      = static_cast<float>(getHeight()) - (kHandleHeight / 2.0f);
 
   auto x = juce::jlimit(xMin, xMax, (float)mPoint.getX());
   auto y = juce::jlimit(yMin, yMax, (float)mPoint.getY());
 
-  pointerPos_ = juce::Point<int>(x, y);
+  pointerPos_ = juce::Point<int>(static_cast<int>(x), static_cast<int>(y));
   audioProcessor.setMixValue(juce::jmap((float)pointerPos_.getY(), yMin, yMax, 1.0f, 0.0f));
 }
